Revert all previous changes to `UnbanBanCommand`.
I'm deciding that we can now go ahead and use the new parts in
the appservice bot.
The reason being that the syntax of old commands
has optional arguments bang in the middle of
the argument list.
Which makes validating and parsing the argument list (not just from
a string, but also from the result of the CommandReader) really hard
and special cased.
When this command does get refactored, the syntax will change.