The error codes in CommandError are not going to work.
They're required which is bs because people are just going to make
up bullshit to satisfy the requirement. And it's rare
that they need to be intercepted.

What's rarer is intercepting them and not being able to change
the implementation.

So what am i going to do instead?

just not have it and you can subtype CommandError.
until i can think of something else.

This commit is NOT contributed under the Apache-2.0 License.
Copyright (C) 2022 Gnuxie <Gnuxie@protonmail.com>
All rights reserved.