Still very WIP on CLIM prompt-for-accept semantics.
So originally I wrote the `ArgumentConsumer` because
I thought that the adaptor would have to increment and call
accept on each stage.

Now I don't know if that makes sense, because obviously that
would be a stateful API and we could just give a callback
for prompts to the argument parser so that it can all be done
within the same continuation, if you like.

Taking this commit as a checkpoint of progress though.