Fix issue where protected rooms could not be removed.
This was introduced in https://github.com/Gnuxie/Draupnir/pull/54/.
Essentially we forgot to remove the room from the protected rooms set,
when the remove command was used.

Ontop of this something to note is that during testing it is clear
that the protected rooms set is loaded when configuring mjolnir,
not when starting it. This is problematic as it means
setup code in `fixtures.ts` does not actually wipe
the protected rooms set.