Remove default banlist command.
In reality this sets a default shortcode,
and for the moment we aren't able to work with shortcodes.

I don't know if it makes sense to bring this command back
in future because we would like to instead have
the ban propagation protection have a setting
for a list to fill client bans directly into.