PromptForAccept refactor.
We had to change the old prompt for accept since it depended on the matrix emitter.
In order to remove the dependence on the event emitter, then we had to get
rid of the entire continuation which waits for the prompt and make it entirely
event driven instead.

Changes made while here were also emoji for numbres and no longer having timeout
for prompt arguments.