Change license to AFL-3.0 (#204)
Please reply with 👍 or approve this PR if you agree to relicense your work under the [Academic Free License](https://opensource.org/license/afl-3-0-php/)

As discussed in #176 we want to change the license to AFL-3.0 so that we have an approved open source license with a condition for distribution or incorporation to provide the source code for the original work (or a link to it and this is implemented currently via the status command. I think we should go a step further and provide the notices file here too as a follow up). 

Contributors previously were asked to agree to provide their [contribution](https://github.com/the-draupnir-project/Draupnir/blob/56e03e9/CONTRIBUTING.md#contributing-code-to-draupnir) under the Apache-2.0 license so that I would be the sole author contributing under the project's [Cooperative Software License](https://github.com/the-draupnir-project/Draupnir/blob/56e03e9/LICENSE), which would in (dubious) theory make relicensing to anything Apache-2.0 compatible straight forward later on. However out of an abundance of caution I would like all the authors who have contributed since forking (to cover the case that someone believes they have contributed under the CSL and not Apache-2.0, and therefore can object) to agree to the license change. This also leaves a concrete paper trail in one place rather than each PR.

Contributors:

- [x] @Mikaela 
- [x] @MTRNord 
- [x] @julianfoad 
- [x] @FSG-Cat 
- [x] @JokerGermany 
- [x] @progval