Make an arbritrary decision when upgrading @types/jsdom.
https://github.com/DefinitelyTyped/DefinitelyTyped#how-do-definitely-typed-package-versions-relate-to-versions-of-the-corresponding-library
Apparently the typings for jsdom are really out of date.
So we have to take the risk and use an older version that
seems to be ok so that we can upgrade the package.

I did consider just using 21.1.6 for both so that we could be
stubbornly accurate but there's no indication how long it will take
for the typings package to be upgraded.