Set `disableStores: true` in matrix-appservice-bridge BridgeOpts.
This stores create database files all over the cwd that use nedb
which shouldn't really be used by the appservice library at this
point.

These files are problematic because they are undocumented to
server admins trying to setup the appservice, and need to be
explicitly accounted for to make permissions work.

The stores enable something called the `BridgeContext` to be
given alongside the event within the `onEvent` handler.
Which we don't use, and it's unclear how it should be used.

The primary motivation for removing the stores is to simplify
work making the appservice available to matrix-docker-ansible-deploy.

We expect to have follow up prs to improve documentation.