Report validation errors.
This has failed miserably because for some reason transform errors
are reported differently to schema validation errors. They aren't
treated like a TypeBox `ValueError`.

`Config.util.extendDeep` also for some obscure reason introduces
`null`s when objects such as `sentry` are undefined,
so the schema had to be changed for that.