Experimental new joiner protection. (#495)
This protection allows room moderators to ban new users from servers (matrix.org). Existing users will be able to participate normally, and if they are accidentally banned, once unbanned rejoining will be unaffected.

This is intended to be used during periods of instability. The protection may be removed entirely or modified in a future release. We will add a documentation page shortly.