Figure out how to apply policies to resident homserver users.
And also do purging deactivation in the background

We introduce "restriction" as a way to refer to an account that has
been suspended or locked or shadowbanned etc.

We also add a UserRestrictionCapability for protections to use the
functionality for suspending users.

We keep an audit log that updates when missing account restrictions
are discovered. (IE those that are prexisting)

The protection that makes these features available is the
HomeserverUserPolicyProtection.