fix Room Discovery protection cleanup (#929)
We have identified that the `RoomDiscovery` component of the `RoomTakedownProtection` and `BlockServersOnInviteProtection` doesn't cleanup properly once disabled. Because a deferred background task used to batch discovered rooms together wasn't being cancelled. We believe this is the cause for the issues described in #927 . Although it remains unclear why this violates the integrity of the hash store in subsequent test runs...

* Update to MPS 3.10.0 for cancellable batches.

* Cancel room discovery batch when protections are disabled.

Possible fix for https://github.com/the-draupnir-project/Draupnir/issues/927.