Fix flakey BlockInvitationsOnServer test case. (#931)
Wait for the new takedown policy to come down `/sync` before testing
the protection can block the takendown user.

There doesn't seem to be a way conceptually to avoid this. I would not
be comfortable with injecting a fake event into the model and we'd
probably have a nightmare creating that concept.

See https://github.com/the-draupnir-project/Draupnir/pull/930#issuecomment-3179386744.