Fix policy room creation in V12 rooms. (#936)
It turns out that we got confused and thought we'd fixed policy room creation when we fixed management room creation.
Even though the PR description never claimed that. In any case it looks like we were not in a very present state of mind while making the change and managed to somehow rely on string comparison for room versions...

Follow up from: https://github.com/the-draupnir-project/Draupnir/pull/924

* Fix subtle bug with management room creation.

I don't really understand why I did this in the first place it was
stupid.  I think morbid curiosity won.

* Update MPS4bot-sdk for V12 policy room creation.

It turns out last time we fixed management room creation but not
policy room creation... and we didn't even do that properly.