Add TimelineRedactionQueue. (#949)
This just batches the redactions going on within a single room to the
same `/messages` request. And blocks the next request for that room
from happening until the current request has completed.

This is necessary because `/messages` pagination is really heavy on
homeservers as it can force them to backfill.

There is also some really strange bug in Synapse that is triggered by
Draupnir's behaviour
https://github.com/element-hq/synapse/issues/18829.